% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{all_miss_col}
\alias{all_miss_col}
\title{Eliminate empty columns of dataset}
\usage{
all_miss_col(dataset, perc)
}
\arguments{
\item{dataset}{A database to summary}

\item{perc}{A numeric value.}
}
\value{
A dataset without empty columns
}
\description{
Eliminate empty columns of dataset
}
