% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{animar}
\alias{animar}
\title{Create animated graphics}
\usage{
animar(data, varA1, two_var, varA2, group, type_graf)
}
\arguments{
\item{data}{A database to use}

\item{varA1}{A character value with the column of x axis}

\item{two_var}{A logical value: T if the graphic contains variable y; F if not}

\item{varA2}{A character value with the column of y axis}

\item{group}{A character value with the column of group}

\item{type_graf}{A character value with the type of graphic}
}
\value{
A plot object
}
\description{
Create animated graphics
}
