% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{cambio_prop}
\alias{cambio_prop}
\title{Change columns class}
\usage{
cambio_prop(dataset, columns, prop)
}
\arguments{
\item{dataset}{A database to use}

\item{columns}{A character vector with the names of columns to change}

\item{prop}{A character value with the class propriety: "factor" or "double"}
}
\value{
A dataset object.
}
\description{
Change columns class
}
