% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{corSig}
\alias{corSig}
\title{Significant correlations}
\usage{
corSig(base, umbrales)
}
\arguments{
\item{base}{A database to use}

\item{umbrales}{A numeric value with Significant umbrals}
}
\value{
Significant correlations
}
\description{
Significant correlations
}
