% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{crear_modelo}
\alias{crear_modelo}
\title{Create model}
\usage{
crear_modelo(
  data,
  var_est,
  var_exp,
  intercep = NULL,
  method,
  percTest,
  ln_vars = F,
  ln_vars_tr = NULL,
  trC_method = "none",
  trC_num,
  trC_rep
)
}
\arguments{
\item{data}{A database to use}

\item{var_est}{A character value with the column of x axis}

\item{var_exp}{A logical value: T if the graphic contains variable y; F if not}

\item{intercep}{A character value with the column of y axis}

\item{method}{A character value with the column of group}

\item{percTest}{A character value with the type of graphic}

\item{ln_vars}{---}

\item{ln_vars_tr}{---}

\item{trC_method}{---}

\item{trC_num}{---}

\item{trC_rep}{---}
}
\value{
A plot object
}
\description{
Create model
}
