% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{imp_data}
\alias{imp_data}
\title{Impute missing values}
\usage{
imp_data(dataset, method)
}
\arguments{
\item{dataset}{A database to use}

\item{method}{A character value with method name: "omit" or "RF"}
}
\value{
A dataset object.
}
\description{
Impute missing values
}
