% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{out_inf_points}
\alias{out_inf_points}
\title{Summary of outliers and influence points}
\usage{
out_inf_points(data, var_est)
}
\arguments{
\item{data}{A database to use}

\item{var_est}{Acharacter value with the column name to study}
}
\value{
A cat summary
}
\description{
Summary of outliers and influence points
}
