% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{redes}
\alias{redes}
\title{Create neural network}
\usage{
redes(
  data,
  var_est,
  vPerc,
  unitsE,
  activationE,
  lDrop = NULL,
  epochsE,
  batch_sizeE,
  validation_splitE
)
}
\arguments{
\item{data}{A database to use}

\item{var_est}{A character value with the column of x axis}

\item{vPerc}{A logical value: T if the graphic contains variable y; F if not}

\item{unitsE}{A character value with the column of y axis}

\item{activationE}{A character value with the column of group}

\item{lDrop}{A character value with the type of graphic}

\item{epochsE}{---}

\item{batch_sizeE}{---}

\item{validation_splitE}{---}
}
\value{
A plot object
}
\description{
Create neural network
}
