% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{Summ_all_data}
\alias{Summ_all_data}
\title{Summary of all data}
\usage{
Summ_all_data(data, var_est = NULL)
}
\arguments{
\item{data}{A database to use}

\item{var_est}{Acharacter value with the column name to study, NULL by default}
}
\value{
A cat summary
}
\description{
Summary of all data
}
