% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruebas.R
\name{Summ_data}
\alias{Summ_data}
\title{Summary of database}
\usage{
Summ_data(dataset)
}
\arguments{
\item{dataset}{A database to summary}
}
\value{
A summary of data base
}
\description{
Summary of database
}
